deployment/app/requirements.txt
flask
opencv-python
torch
numpy
Pillow
ultralytics